Move features to their own package
