+++ android/libraries/rib-base/src/main/java/com/badoo/ribs/core/routing/backstack/feature/BackStackFeature.kt
@@ -1,4 +1,4 @@
-package com.badoo.ribs.core.routing.backstack
+package com.badoo.ribs.core.routing.backstack.feature
 
 import android.os.Parcelable
 import com.badoo.mvicore.element.Actor
@@ -6,16 +6,16 @@ import com.badoo.mvicore.element.Bootstrapper
 import com.badoo.mvicore.element.Reducer
 import com.badoo.mvicore.element.TimeCapsule
 import com.badoo.mvicore.feature.BaseFeature
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.Action
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.Action.Execute
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.Effect
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.State
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.Wish
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.Wish.NewRoot
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.Wish.Pop
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.Wish.Push
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.Wish.PushOverlay
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.Wish.Replace
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.Action
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.Action.Execute
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.Effect
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.State
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.Wish
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.Wish.NewRoot
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.Wish.Pop
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.Wish.Push
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.Wish.PushOverlay
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.Wish.Replace
 import io.reactivex.Observable
 import io.reactivex.Observable.empty
 import io.reactivex.Observable.just
