+++ android/libraries/rib-base/src/test/java/com/badoo/ribs/core/routing/backstack/BackStackManagerTest.kt
@@ -3,12 +3,13 @@ package com.badoo.ribs.core.routing.backstack
 import com.badoo.mvicore.element.TimeCapsule
 import com.badoo.ribs.core.helper.TestRouter
 import com.badoo.ribs.core.helper.TestRouter.Configuration
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.State
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.Wish.NewRoot
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.Wish.Pop
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.Wish.Push
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.Wish.PushOverlay
-import com.badoo.ribs.core.routing.backstack.BackStackFeature.Wish.Replace
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.State
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.Wish.NewRoot
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.Wish.Pop
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.Wish.Push
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.Wish.PushOverlay
+import com.badoo.ribs.core.routing.backstack.feature.BackStackFeature.Wish.Replace
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.mock
 import org.junit.Assert.assertEquals
@@ -24,7 +25,7 @@ class BackStackManagerTest {
     private lateinit var timeCapsuleEmpty: TimeCapsule<BackStackFeature.State<TestRouter.Configuration>>
     private lateinit var timeCapsuleWithContent: TimeCapsule<BackStackFeature.State<TestRouter.Configuration>>
     private lateinit var backstackInTimeCapsule: List<Configuration>
-    private lateinit var backStackManager: BackStackFeature<TestRouter.Configuration>
+    private lateinit var backStackManager: BackStackFeature<Configuration>
 
     @Before
     fun setUp() {
